Rearrange classes to bust up package cycle.
