+++ subprojects/launcher/src/main/java/org/gradle/tooling/internal/provider/action/BuildModelAction.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013 the original author or authors.
+ * Copyright 2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,22 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-package org.gradle.tooling.internal.provider;
+package org.gradle.tooling.internal.provider.action;
 
 import org.gradle.api.internal.StartParameterInternal;
-import org.gradle.tooling.internal.provider.serialization.SerializedPayload;
 import org.gradle.internal.build.event.BuildEventSubscriptions;
+import org.gradle.tooling.internal.protocol.ModelIdentifier;
 
-public class ClientProvidedBuildAction extends SubscribableBuildAction {
+public class BuildModelAction extends SubscribableBuildAction {
     private final StartParameterInternal startParameter;
-    private final SerializedPayload action;
+    private final String modelName;
     private final boolean runTasks;
 
-    public ClientProvidedBuildAction(StartParameterInternal startParameter, SerializedPayload action, boolean runTasks, BuildEventSubscriptions clientSubscriptions) {
+    public BuildModelAction(StartParameterInternal startParameter, String modelName, boolean runTasks, BuildEventSubscriptions clientSubscriptions) {
         super(clientSubscriptions);
         this.startParameter = startParameter;
-        this.action = action;
+        this.modelName = modelName;
         this.runTasks = runTasks;
     }
 
@@ -37,8 +36,8 @@ public class ClientProvidedBuildAction extends SubscribableBuildAction {
         return startParameter;
     }
 
-    public SerializedPayload getAction() {
-        return action;
+    public String getModelName() {
+        return modelName;
     }
 
     @Override
@@ -48,6 +47,6 @@ public class ClientProvidedBuildAction extends SubscribableBuildAction {
 
     @Override
     public boolean isCreateModel() {
-        return true;
+        return !ModelIdentifier.NULL_MODEL.equals(modelName);
     }
 }
