+++ subprojects/launcher/src/main/java/org/gradle/tooling/internal/provider/action/TestExecutionRequestAction.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 the original author or authors.
+ * Copyright 2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.tooling.internal.provider;
+package org.gradle.tooling.internal.provider.action;
 
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
