+++ subprojects/launcher/src/test/groovy/org/gradle/tooling/internal/provider/action/BuildActionSerializerTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 the original author or authors.
+ * Copyright 2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,19 +14,13 @@
  * limitations under the License.
  */
 
-package org.gradle.launcher.cli
+package org.gradle.tooling.internal.provider.action
 
 import org.gradle.api.internal.StartParameterInternal
 import org.gradle.internal.build.event.BuildEventSubscriptions
 import org.gradle.internal.serialize.SerializerSpec
-import org.gradle.launcher.cli.action.BuildActionSerializer
-import org.gradle.launcher.cli.action.ExecuteBuildAction
 import org.gradle.tooling.events.OperationType
 import org.gradle.tooling.events.test.internal.DefaultDebugOptions
-import org.gradle.tooling.internal.provider.BuildModelAction
-import org.gradle.tooling.internal.provider.ClientProvidedBuildAction
-import org.gradle.tooling.internal.provider.ClientProvidedPhasedAction
-import org.gradle.tooling.internal.provider.TestExecutionRequestAction
 import org.gradle.tooling.internal.provider.serialization.SerializedPayload
 import spock.lang.Unroll
